Time: 56 ms (50.80%), Space: 6.2 MB (76.00%) - LeetHub

+++ 1155-number-of-dice-rolls-with-target-sum/1155-number-of-dice-rolls-with-target-sum.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int numRollsToTarget(int n, int k, int target) {
         int mod = 1e9+7;
-        long dp[31][2001] = {0};
+        int dp[31][2001] = {0};
         for(int i = 0; i < k; i++){
             dp[0][i] = 1;
         }
